Add support for anonymous objects directly in route properties response.
