+++ packages/core/src/lib/generators/ServiceGenerator.ts
@@ -79,7 +79,8 @@ export const getRequestContractType = (
 };
 
 const getContractResult = (
-  endpointDescription: EndpointDescription
+  endpointDescription: EndpointDescription,
+  swagger: SwaggerSchema
 ): Maybe<string> => {
   const getSchemas = (operation: Operation) =>
     Object.entries(operation.responses).map((e) => ({
@@ -90,12 +91,12 @@ const getContractResult = (
   const getTypeName = (schema: Schema, isArray: boolean) => {
     if (schema.oneOf) {
       const typeNames = schema.oneOf
-        .map((s) => getTypeNameFromSchema(s))
+        .map((s) => getTypeNameFromSchema(s, swagger))
         .join(" | ");
       return isArray ? `(${typeNames})[]` : typeNames;
     }
 
-    const typeName = getTypeNameFromSchema(schema);
+    const typeName = getTypeNameFromSchema(schema, swagger);
     return isArray ? `${typeName}[]` : typeName;
   };
 
@@ -439,7 +440,7 @@ export const generateServices = (swagger: SwaggerSchema) => {
       });
 
       const contractResult =
-        getContractResult(endpointDescription).orDefault("any");
+        getContractResult(endpointDescription, swagger).orDefault("any");
 
       if (
         endpointDescription.methodType === "POST" ||
